refactor(checker): split instantiated-property-access display out of display_formatting.rs to hold 2000 LOC
display_formatting.rs sits at exactly 2000 physical lines on main, so any touch
trips the hard 2000-line checker-source arch cap (it gates #15447/#15449).
Extract instantiated_property_access_annotation_display into a new sibling
display_formatting_property_access.rs (pub(super)); its three private helper
deps (collect_type_param_display_replacements,
later_literal_argument_replacement_for_type_param,
raw_callback_return_type_param_name) are widened to pub(super) so the moved
method can still reach them. Pure code motion; no behavior change.

display_formatting.rs 2000->1628; new sibling 385. No other checker src file
affected. Hand-verified: both call sites resolve via pub(super), imports
complete, NodeAccess still used in the parent (2 calls) and not needed in the
sibling.
